refactor: migrate `mkdocstrings` MkDocs plugin replacement
