+++ crates/zensical/src/compat/mkdocs.rs
@@ -5,4 +5,5 @@
 
 //! MkDocs compatibility modules.
 
+pub mod mkdocstrings;
 pub mod search;
